% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplotty.R
\name{myplotty}
\alias{myplotty}
\title{myplotty function}
\usage{
myplotty(value, dataset)
}
\arguments{
\item{value}{A numeric vector of data values (e.g., a column from a dataset).
This represents the values for which Z-scores are to be computed.}

\item{dataset}{A numeric vector containing the dataset for calculating the Z-scores.
The mean and standard deviation of this dataset are used to compute the Z-scores.}
}
\value{
A numeric vector of Z-scores corresponding to each value in the `value` vector.
The Z-score is calculated as: (value - mean(dataset)) / sd(dataset).
}
\description{
myplotty function
}
\examples{
value <- 2
spruce <- c(1, 2, 3, 4, 5, 6)
myplotty(value, spruce)

}
